move out DevTools from App to simplify App
